perf(web): kill sidebar prefetch + lazy-load command palette dialog
Production measurement showed two concrete waste sources on every
dashboard mount:

1. Sidebar fires 18 sibling-page RSC requests as Next.js Link prefetches
   them on viewport. Each costs ~327ms in the web middleware (Supabase
   auth + org_members lookups), with ~17% returning 503 once Vercel's
   per-instance concurrency cap kicks in. None of those prefetches
   affect the dashboard the user actually sees.

2. The cmdk-based CommandPalette dialog ships in the initial dashboard
   chunk even though 99% of dashboard mounts never open it.

Fixes:

  apps/web/components/layout/sidebar.tsx
    All sidebar Links now use prefetch={false} unconditionally. Removed
    the linkPrefetchFor() import since heavy/light distinction no longer
    matters at this surface. KpiCard + the inline drill-down Links in
    dashboard-client.tsx still call linkPrefetchFor() — heavy-pages.ts
    is unchanged.
    Trade-off: first click to a sidebar page now pays ~300-500ms cold
    cost. Acceptable because (a) users who navigate actively warm the
    paths they care about via subsequent hover/click, and (b) the cost
    was previously absorbed *by every dashboard mount*, paid whether
    they used it or not.

  apps/web/components/command-palette.tsx
    Extracted the cmdk-heavy CommandPaletteDialog into a sibling file
    and dynamic-imported it. Conditional render (`{open && <Dialog />}`)
    so the chunk only fetches on the first ⌘K press, not on dashboard
    mount.
    The Provider, context, and keyboard listener stay in the eager
    chunk (lightweight, no cmdk).

  apps/web/components/command-palette-dialog.tsx (new)
    Hosts the cmdk imports + Dialog UI. Receives navGroups as props
    so the call site can stay generic.

Expected impact:

  Sidebar prefetch RSC count       18 → 0
  503 burst on dashboard mount     3-4/mount → 0
  Initial dashboard JS chunk       ~30-50KB lighter (cmdk + Command UI deferred)
  First ⌘K press                   one-time ~50ms chunk fetch

No expected impact on dashboard cold/warm baseline since those numbers
are dominated by /api/v1/* request latency, not prefetch noise.

Verified locally: typecheck + lint pass.

Refs: production measurement after Sprint 1+2 + PR #117 — RSC 327ms
median across sibling pages, 342KB largest JS chunk.
